Updated the bash scripts

+++ scripts/eclipseSetup.sh
@@ -1,6 +1,10 @@
 cd ..
 cd ..
-rm -r Eclipse
+if [ -d $Eclipse];
+then
+	echo "Removing Folder: Eclipse"
+	rm -r Eclipse
+fi
 cd Eclipse
 cmake -G "Eclipse CDT4 - Unix Makefiles" ../MusicGame
 cd ../MusciGame
