+++ scripts/setupBuild.sh
@@ -1,5 +1,8 @@
 cd ..
-rm -r build
+if [ -d $build]
+then
+ rm -r build
+fi
 mkdir build
 cd build
 cmake ..
